fix: use int album id from MediaStore

+++ android/app/src/main/kotlin/com/example/musicplayer/MainActivity.kt
@@ -87,7 +87,7 @@ class MainActivity : AudioServiceActivity() {
             val durationCol = cursor.getColumnIndexOrThrow(MediaStore.Audio.Media.DURATION)
 
             while (cursor.moveToNext()) {
-                val albumId = cursor.getLong(albumIdCol)
+                val albumId = cursor.getInt(albumIdCol)
 
                 songs.add(
                     mapOf(
